Hiding the RxActivityResult implementation behind a class not prefixed with Rx so that the generateKotlin task doesn't choke.
